+++ R/accessor.R
@@ -1,4 +1,4 @@
-#' Sample rates.
+#' Accessor functions
 #'
 #' This is a convenience function for those cases where the comtrade file has
 #' **exactly one** sampling frequency.
@@ -14,7 +14,7 @@
 #'   library("ieeecomtrade")
 #'
 #'   ct_sample_per_cycle(keating_1999)
-#'   ct_sample_rate_nominal(keating_1999)
+#'   ct_sampling_frequency_nominal(keating_1999)
 #' @export
 #'
 ct_sample_per_cycle <- function(ct) {
@@ -39,11 +39,45 @@ ct_sample_per_cycle <- function(ct) {
 #' @export
 #' @return A `numeric`, estimate of nominal sampling-rate (Hz.).
 #'
-ct_sample_rate_nominal <- function(ct) {
+ct_sampling_frequency_nominal <- function(ct) {
 
-  sample_rate_nominal <-
+  assertthat::assert_that(
+    inherits(ct, "comtrade")
+  )
+
+  sampling_frequency_nominal <-
     ct_sample_per_cycle(ct) *
     ct$config$lf
 
-  sample_rate_nominal
+  sampling_frequency_nominal
 }
+
+#' @rdname ct_sample_per_cycle
+#' @export
+#' @return A `numeric`, electrical frequency (Hz.)
+#'
+ct_electrical_frequency_nominal <- function(ct) {
+
+  assertthat::assert_that(
+    inherits(ct, "comtrade")
+  )
+
+    # may have to make sure this rounds to 60 or 50
+
+  ct$config$lf
+}
+
+#' @rdname ct_sample_per_cycle
+#' @export
+#' @return A `numeric`, electrical frequency (Hz.)
+#'
+ct_sampling_frequency <- function(ct) {
+
+  assertthat::assert_that(
+    inherits(ct, "comtrade"),
+    identical(ct$config$nrates, 1L)
+  )
+
+  ct$config$sampling_rate$samp
+}
+